requirements.txt
bcrypt==4.1.2
black==24.1.1       
cffi==1.16.0        
click==8.1.7        
colorama==0.4.6     
cryptography==42.0.0
exceptiongroup==1.2.0
iniconfig==2.0.0
mypy-extensions==1.0.0
packaging==23.2
paramiko==3.4.0
pathspec==0.12.1
phoenix6==24.1.0
photonlibpy==2024.2.3
Pint==0.23
platformdirs==4.2.0
pluggy==1.3.0
pycparser==2.21
pyfrc==2024.0.1
pyfrc==2024.0.1
PyNaCl==1.5.0
pynetconsole==2.0.4
pyntcore==2024.2.1.1
pytest==7.4.4
pytest-reraise==2.1.2
robotpy==2024.2.1.0
robotpy-apriltag==2024.2.1.1
robotpy-cli==2024.0.0
robotpy-commands-v2==2024.0.0b4
robotpy-cscore==2024.2.1.1
robotpy-ctre==2024.1.1
robotpy-hal==2024.2.1.1
robotpy-halsim-ds-socket==2024.2.1.1
robotpy-halsim-gui==2024.2.1.1
robotpy-halsim-ws==2024.2.1.1
robotpy-installer==2024.1.3
robotpy-navx==2024.1.0
robotpy-pathplannerlib==2024.1.3
robotpy-playingwithfusion==2024.0.0
robotpy-rev==2024.2.0
robotpy-romi==2024.2.1.1
robotpy-wpilib-utilities==2024.0.0
robotpy-wpimath==2024.2.1.1
robotpy-wpinet==2024.2.1.1
robotpy-wpiutil==2024.2.1.1
robotpy-xrp==2024.2.1.1
tomli==2.0.1
typing_extensions==4.9.0
wpilib==2024.2.1.1